Semaphore file is not written in case of error
-> parsing errors
-> if execution of simulation command failed

(also changed data type of value returned by function genmain)

+++ include/genesis.h
@@ -9,7 +9,5 @@
 
 using namespace std;
 
-
-
-double genmain(string, string, string, int,bool);
+int genmain(string, string, string, int, bool);
 #endif
